a few changes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
-            $table->tinyInteger('role_id');
+            $table->integer('role_id');
             $table->string('password');
             $table->rememberToken();
             $table->timestamp('email_verified_at')->nullable();
