+++ database/migrations/2022_02_11_133540_create_books_table.php
@@ -17,7 +17,7 @@ class CreateBooksTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('bookcase');
-            $table->tinyInteger('user_id');
+            $table->integer('user_id');
             // $table->string('section');
             // $table->string('category');
             $table->string('author');
